+++ R/dupree.R
@@ -52,23 +52,24 @@
 #' @param        files         A set of files over which code-duplication
 #'   should be measured.
 #'
-#' @param        min_block_size   `dupree` uses a notion of non-trivial
+#' @param        min_block_size   \code{dupree} uses a notion of non-trivial
 #'   symbols.  These are the symbols / code-words that remain after filtering
-#'   out really common symbols like `<-`, `,`, etc. After filtering out these
-#'   symbols from each code-block, only those blocks containing at least
-#'   `min_block_size` symbols are used in the inter-block code-duplication
-#'   measurement.
+#'   out really common symbols like \code{<-}, \code{,}, etc. After filtering
+#'   out these symbols from each code-block, only those blocks containing at
+#'   least \code{min_block_size} symbols are used in the inter-block
+#'   code-duplication measurement.
 #'
 #' @param        ...           Unused at present.
 #'
-#' @return        A tibble. Each row in the table summarises the comparison
-#' between two code-blocks (block 'a' and block 'b') in the input files. Each
-#' code-block in the pair is indicated by: i) the file (file_a / file_b) that
-#' contains it; ii) its position within that file (block_a / block_b; 1 being
-#' the first code-block in a given file); and iii) the line where that
-#' code-block starts in that file (line_a / line_b). The pairs of code-blocks
-#' are ordered by decreasing similarity. Any match that is returned is either
-#' the top hit for block 'a' or for block 'b' (or both).
+#' @return        A \code{tibble}. Each row in the table summarises the
+#'   comparison between two code-blocks (block 'a' and block 'b') in the input
+#'   files. Each code-block in the pair is indicated by: i) the file
+#'   (\code{file_a} / \code{file_b}) that contains it; ii) its position within
+#'   that file (\code{block_a} / \code{block_b}; 1 being the first code-block in
+#'   a given file); and iii) the line where that code-block starts in that file
+#'   (\code{line_a} / \code{line_b}). The pairs of code-blocks are ordered by
+#'   decreasing similarity. Any match that is returned is either the top hit for
+#'   block 'a' or for block 'b' (or both).
 #'
 #' @importFrom   magrittr      %>%
 #'
@@ -87,43 +88,47 @@
 #' # `min_block_size` is too large, none of the code-blocks will be analysed
 #' # and the results will be empty:
 #' dupree(example_file, min_block_size = 40)
-#'
 #' @export
 
-dupree <- function(files, min_block_size = 20, ...) {
+dupree <- function(files, min_block_size = 40, ...) {
   preprocess_code_blocks(files, min_block_size) %>%
-    find_best_matches()
+    find_best_matches() %>%
+    as_dups()
 }
 
 ###############################################################################
 
-#' `dupree_dir` - run duplicate-code detection over all R-files in a directory
+#' Run duplicate-code detection over all R-files in a directory
 #'
 #' @inheritParams   dupree
 #'
-#' @param        path          A directory. All files in this directory that
-#' have a ".R", ".r" or ".Rmd" extension will be checked for code duplication.
+#' @param        path          A directory (By default the current working
+#'   directory). All files in this directory that have a ".R", ".r" or ".Rmd"
+#'   extension will be checked for code duplication.
 #'
-#' @param        filter        A pattern for use in grep - this is used to
-#' keep only particular files: eg, filter = "classes" would compare files with
-#' `classes` in the filename
+#' @param        filter        A pattern for use in grep - this is used to keep
+#'   only particular files: eg, filter = "classes" would compare files with
+#'   `classes` in the filename
 #'
-#' @param        ...           Further arguments for grep. For example,
-#' `filter = "test", invert = TRUE` would disregard all files with `test` in
-#' the file-path.
+#' @param        ...           Further arguments for grep. For example, `filter
+#'   = "test", invert = TRUE` would disregard all files with `test` in the
+#'   file-path.
 #'
 #' @param       recursive     Should we consider files in subdirectories as
-#' well?
+#'   well?
 #'
 #' @seealso     dupree
 #'
 #' @export
 
-dupree_dir <- function(path,
-                       min_block_size = 20,
+dupree_dir <- function(path = ".",
+                       min_block_size = 40,
                        filter = NULL,
                        ...,
                        recursive = TRUE) {
+  if (!dir.exists(path)) {
+    stop("The path ", path, " does not exist")
+  }
   files <- dir(
     path,
     pattern = ".*(.R|.r|.Rmd)$", full.names = TRUE, recursive = recursive
@@ -133,28 +138,40 @@ dupree_dir <- function(path,
   } else {
     files[grep(pattern = filter, x = files, ...)]
   }
+
   dupree(keep_files, min_block_size)
 }
 
 ###############################################################################
 
-#' `dupree_package` - run duplicate-code detection over all files in a
-#' package's `R` directory
+#' Run duplicate-code detection over all files in the `R` directory of a
+#' package
+#'
+#' The function fails if the path does not look like a typical R package (it
+#' should have both an R/ subdirectory and a DESCRIPTION file present).
 #'
 #' @inheritParams   dupree
 #'
 #' @param        package       The name or path to the package that is to be
-#' checked.
+#'   checked (By default the current working directory).
 #'
 #' @seealso      dupree
 #'
+#' @include      utils.R
 #' @export
 
-dupree_package <- function(package,
-                           min_block_size = 20) {
-  # nolint start
+dupree_package <- function(package = ".",
+                           min_block_size = 40) {
+  if (!dir.exists(package)) {
+    stop("The path ", package, " does not exist")
+  }
+  if (!has_description(package)) {
+    stop("The path ", package, " is not an R package (no DESCRIPTION)")
+  }
+  if (!has_r_source_dir(package)) {
+    stop("The path", package, " is not an R package (no R/ subdir)")
+  }
   dupree_dir(package, min_block_size, filter = paste0(package, "/R/"))
-  # nolint end
 }
 
 ###############################################################################